[parkhj0423] 일부 DTO 파싱시 오류나서 Entity로 변환 못하던 현상 해결, 이미지 없는 Entity 대체 이미지 지정

+++ Beer-Recommendation/DataLayer/DTO/BeerDTO.swift
@@ -14,9 +14,9 @@ struct BeerDTO : Codable {
     let image_url: String?
     let abv: Double?
     let ibu: Double?
-    let target_fg: Int?
+    let target_fg: Double?
     let target_og: Double?
-    let ebc: Int?
+    let ebc: Double?
     let srm, ph: Double?
     let attenuation_level: Double?
     let volume, boil_volume: BoilVolumeDTO?
