+++ Beer-Recommendation/DomainLayer/Entity/BeerEntity.swift
@@ -17,11 +17,11 @@ struct BeerEntity : Codable, Identifiable {
     // 쓴맛의 정도 숫자가 높을 수록 크다 (보통의 IPA가 70 정도)
     let ibu: Double?
     // 효모를 넣고 발효 후 물에 비해 얼마나 무거운 가를 나타내는 수치
-    let targetFg: Int?
+    let targetFg: Double?
     // 발효 전 물에 비해 얼마나 무거운 가를 나타내는 수치
     let targetOg: Double?
     // 유럽 기준) 색의 짙은 정도 숫자가 높을 수록 짙다 (보통의 IPA가 12~13, 스타우트가 35)
-    let ebc: Int?
+    let ebc: Double?
     // 미국 기준) 색의 짙은 정도 숫자가 높을 수록 짙다 (보통의 IPA가 12~13, 스타우트가 35)
     let srm : Double?
     let ph : Double?
